# docs/openapi.yaml
openapi: 3.1.0
info:
  title: UPP Gateway API
  version: 0.1.0
  description: |
    Universal Prediction Protocol Gateway API provides unified access to prediction markets across multiple providers via a single API.

    ## Authentication

    The API uses two authentication methods:
    - **ApiKeyAuth**: X-API-Key header for API key-based authentication
    - **BearerAuth**: Bearer token in Authorization header for JWT-based authentication

    Protected endpoints require authentication. Public endpoints are rate-limited but do not require credentials.

    ## Rate Limiting

    All endpoints are rate-limited. Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests per minute
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `Retry-After`: Seconds to wait before retrying (on 429 responses)

  contact:
    name: UPP Team
    url: https://github.com/upp-labs
    email: support@upp.dev
  license:
    name: Apache-2.0
    url: https://opensource.org/licenses/Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://gateway:8080
    description: Docker containerized gateway
  - url: https://api.upp.dev
    description: Production server
...
paths:
...
  /upp/v1/ws:
    get:
      operationId: websocketUpgrade
      summary: WebSocket upgrade
      description: Upgrade to WebSocket connection for real-time data streaming
      tags: [WebSocket]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '101':
          description: WebSocket connection established
        '400':
          description: Invalid WebSocket upgrade request
        '401':
          description: Unauthorized
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication using X-API-Key header

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token authentication using JWT